requirements.txt
torch
adam-atan2
einops
tqdm
coolname
pydantic
argdantic
wandb
omegaconf
hydra-core
huggingface_hub
packaging 
ninja 
wheel 
setuptools 
setuptools-scm
pydantic-core
huggingface_hub 
numba
triton
sentencepiece
zstandard
huggingface_hub[cli]
